# src/Web_Development/Recipes_Backend/API/API.yaml
openapi: 3.0.0
info:
  title: Recipes Project API
  description: This is a API for the project assginment in web development environments
  contact:
    email: attiasas@post.bgu.ac.il
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.1
servers:
- url: http://localhost:3000
  description: SwaggerHub API Auto Mocking
...
paths:
...
  /recipes/recipe/{recipeId}/preview:
    get:
      tags:
      - Recipes
      summary: Get a preview of recipe base on ID
      description: Get a preview of recipe base on ID
      parameters:
      - name: recipeId
        in: path
        description: the id of the recipe
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: recipe information returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipePreview'
        "400":
          description: Not Found
  /recipes/recipe/{recipeId}:
    get:
      tags:
      - Recipes
      summary: Get the full information of the recipe
      description: Get all the information of the recipe by his id
      parameters:
      - name: recipeId
        in: path
        description: the id of the recipe
        required: true
        style: simple
        explode: false
        schema:
          type: integer
      responses:
        "200":
          description: recipe information returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        "400":
          description: illegal parameters
  /recipes/search/query/{recipeName}/amount/{numberOfResults}:
    get:
      tags:
      - Recipes
      summary: search for recipes based on parameters
      description: search for recipes from the server data base
      parameters:
      - name: numberOfResults
        in: path
        description: number of returned recipes
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          default: 5
          enum:
          - 5
          - 10
          - 15
      - name: recipeName
        in: path
        description: name of a recipe
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: cuisine
        in: query
        description: the cuisine that the searched recipe belongs to
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: diet
        in: query
        description: the diet that the searched recipe belongs to
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: intolerances
        in: query
        description: the intolerances that the searched recipe belongs to
        required: false
        style: form
        explode: true
        schema:
          type: string
      responses:
        "200":
          description: recipes returned
          content:
            application/json:
              schema:
                type: array
                description: list of results
                items:
                  $ref: '#/components/schemas/RecipePreview'
        "204":
          description: No recipes found for the search
        "400":
          description: illegal parameters
  /recipes/fetchrandom:
    get:
      tags:
      - Recipes
      summary: Returns three random recipes
      description: fetch three random recipes from the server database
      parameters:
      - name: number
        in: query
        description: the max amount of recipes that will return
        required: true
        style: form
        explode: true
        schema:
          type: integer
      responses:
        "200":
          description: Information on three recipes retured successfully
          content:
            application/json:
              schema:
                type: array
                description: list of results
                items:
                  $ref: '#/components/schemas/RecipePreview'
components:
  schemas:
...
    RecipePreview:
      type: object
      properties:
        recipeId:
          type: integer
          description: the id of the recipe
          example: 123
        name:
          type: string
          description: the name of the recipe
          example: Pasta
        image:
          pattern: URL
          type: string
          description: the URL of the recipe final result image
          example: https://upload.wikimedia.org/wikipedia/commons/1/19/Pac-Man.svg
        readyInMinutes:
          pattern: double
          type: number
          description: the time in minutes that the recipe takes
          example: 2.5
        likes:
          type: integer
          description: the amount of likes that this recipe recived
          example: 132
        vegetarian:
          type: boolean
          description: is this recipe for vegetarian
          example: true
        vegan:
          type: boolean
          description: is this recipe for vegans
          example: false
        glutenFree:
          type: boolean
          description: is this recipe contains gluten
          example: true
    Recipe:
      type: object
      properties:
        recipeId:
          type: integer
          description: the id of the recipe
          example: 123
        name:
          type: string
          description: the name of the recipe
          example: Pasta
        image:
          pattern: URL
          type: string
          description: the URL of the recipe final result image
          example: https://upload.wikimedia.org/wikipedia/commons/1/19/Pac-Man.svg
        readyInMinutes:
          pattern: double
          type: number
          description: the time in minutes that the recipe takes
          example: 2.5
        likes:
          type: integer
          description: the amount of likes that this recipe recived
          example: 132
        vegetarian:
          type: boolean
          description: is this recipe for vegetarian
          example: true
        vegan:
          type: boolean
          description: is this recipe for vegans
          example: false
        glutenFree:
          type: boolean
          description: is this recipe contains gluten
          example: true
        servings:
          type: integer
          description: the amount of servings that this recipe provied
          example: 3
        ingredients:
          type: array
          description: the ingredients that this recipe requires
          items:
            $ref: '#/components/schemas/Ingredient'
        steps:
          type: array
          description: the steps for preparing this recipe
          items:
            $ref: '#/components/schemas/Steps'
...
    Ingredient:
      type: object
      properties:
        unit:
          type: string
          description: the type of unit of the metric
          example: Tbsps
        amount:
          pattern: double
          type: number
          description: the value of the amount in the given units
          example: 2
        name:
          type: string
          description: the name of the ingredient
          example: flour
        image:
          type: string
          description: the image URL of the ingredient
          example: https://upload.wikimedia.org/wikipedia/commons/1/19/Pac-Man.svg
    Steps:
      type: object
      properties:
        step_number:
          type: integer
          description: step number in the recipe
          example: 12
        step_content:
          type: string
          description: step in the recipe
          example: Preheat the oven to 200 degrees F.
      description: steps in the recipe
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      name: session
      in: cookie